// generated/src/app/api-models/group-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class GroupMaster {
  seqGroupId: number;
  groupId: string;
  shortName: string;
  levelCode: string;
  seqParentId: number;
  groupName1: string;
  groupName2: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  groupType: string;
  sicCode: string;
  userDefined1: string;
  userDefined2: string;
  prorationMethod: string;
  billingLevel: string;
  billingCycle: string;
  billingBatch: string;
  noOfRetroMonths: number;
  billedThroughDate: Date;
  gracePeriod: number;
  sortOption: string;
  nextOpenStartDt: Date;
  nextOpenEndDate: Date;
  holdReason: string;
  holdDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  paidThroughDate: Date;
  country: string;
  billingFrequency: string;
  commonBillingDate: Date;
  paymentDueDaysAfter: number;
  rateFreezeCalc: string;
  invoicePrintFormat: string;
  useEftFlg: string;
  adultDepAgeParam: number;
  maxDependentsCharge: number;
  calcDueDateRule: string;
  calcDueDateDays: number;
  ptdUseReasonCde: string;
  ptdThresholdPct: number;
  nationalEmployerId: string;
  taxId: string;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  userDefined6: string;
  userDefined7: string;
  userDefined8: string;
  userDefined9: string;
  userDefined10: string;
  userDefined11: string;
  userDefined12: string;
  userDefined13: string;
  userDefined14: string;
  userDefined15: string;
  userDefined16: string;
  userDefined17: string;
  userDefined18: string;
  userDefined19: string;
  userDefined20: string;
  userDefined21: string;
  userDefined22: string;
  userDate3: Date;
  userDate4: Date;
  userDate5: Date;
  userDate6: Date;
  userDate7: Date;
  userDate8: Date;
  userDate9: Date;
  userDate10: Date;
  userDate11: Date;
  userDate12: Date;
  userDate13: Date;
  userDate14: Date;
  userDate15: Date;
  userDate16: Date;
  userDate17: Date;
  userDate18: Date;
  userDate19: Date;
  userDate20: Date;
  userDate21: Date;
  userDate22: Date;
  comments: string;
}